refactor(workstation): extract status + conflicts surfaces from inkRuntime (#890 phase 5a.3)
Third batch of surface extractions, building on #898 (5a.2). The two
status-related surfaces are paired (status delegates to conflicts when
an in-progress merge / rebase / cherry-pick / revert has unresolved
files) and self-contained — clean cuts.

Surfaces moved (2):
  - status     → src/workstation/surfaces/status/index.ts        (198)
  - conflicts  → src/workstation/surfaces/conflicts/index.ts     (135)

Helpers moved with renderStatusSurface (used only by it):
  - StatusSurfaceRow type
  - buildStatusSurfaceRows
  - capitalizeGroupName
  - formatStatusFilterMask
  - isStatusFilterMaskActive

inkRuntime.ts: 4,999 → 4,722 LOC (-277 this PR; -1,317 cumulative since
5a started, -22% off the original 6,039). 6 imports pruned that became
unused after extraction:
  STAGE_STATUS_DOT, getStageStatusDotColor (chrome/iconography),
  formatLogInkStatusEmpty (chrome/surfaceStates),
  LogInkStatusFilterMask (commands/log/inkViewModel),
  WorktreeFile, WorktreeFileGroup (git/statusData).
applyStatusFilterMask + groupWorktreeFiles stay — still used by
LogInkApp's effect hooks for input-context rectification.

10 of 12 main surfaces are now in their own modules. 5a.4-5a.6 cover
the remaining three: history (renderHistoryPanel + supporting), diff
(309-LOC renderDiffSurface + the local split-diff helpers), and detail
(renderDetailPanel + inspector + commit-diff-detail + compose-context
+ file-list + 4 preview panels).

Validation: tsc clean, lint clean, 680/680 suites & 6571/6571 tests
passing (same baseline as post-5a.2). No behavior change — both
surfaces render byte-identically before/after.
